server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
reportlab==4.0.7
tenacity==8.2.3
openai==1.3.7
python-dotenv==1.0.0
watchdog==3.0.0
